requirements.txt
flask==3.0.0
supabase==2.28.3
python-dotenv==1.2.2
